feat: Display `Collegiate` response with `rich`
This is not fully complete, but works with most words. Bugs will be ironed out in the future.

This is heavily coupled to the configuration system, and should ideally allow users to change text colors for key
display elements. As the `display` module changes, so should the `default_config.toml` file.